requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Edge Impulse SDK
edge-impulse-linux>=1.0.0

# Data processing
tqdm>=4.62.0
pyyaml>=5.4.0
requests>=2.28.0

# Image augmentation (optional, Edge Impulse has built-in)
albumentations>=1.1.0

# Kaggle API (optional, for dataset download)
kaggle>=1.5.12

# Utilities
python-dotenv>=0.19.0

